+++ doc/release-notes.md
@@ -4,4 +4,7 @@ Bitcoin ABC version 0.27.6 is now available from:
 
   <https://download.bitcoinabc.org/0.27.6/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features and fixes:
+ - The `ancestorcount`, `ancestorsize` and `ancestorfees` fields from the
+   `listunspent` RPC have been removed. They were deprecated since the v0.27.0
+   and are no longer computed since the May 15th, 2023 eCash network upgrade.